Reorganization of all files of the repo. Use of relative import based on file context
