+++ Examples/ESSeisInversionDriver.py
@@ -1,41 +1,40 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Nov 20 18:19:51 2020
+Created on Fri Nov 20 18:19:26 2020
 
 @author: dariograna
 
 Reference: Grana and de Figueiredo, 2021, SeReMpy
 """
 
-#% Ensemble Smoother Petrophysical Inversion Driver %%
+#% Ensemble Smoother Seismic Inversion Driver %%
 # In this script we apply the Ensemble Smoother inversion method
-# (Liu and Grana, 2018) to predict the petrophysical properties 
-# from seismic data.
+# (Liu and Grana, 2018) to predict the elastic properties (P- and S-wave
+# velocity and density) from seismic data.
 
 from scipy.io import loadmat
 import matplotlib.pyplot as plt
 from scipy import signal
 import numpy as np
-from Inversion import *
-from Geostats import CorrelatedSimulation
-from RockPhysics import LinearizedRockPhysicsModel
+
+from context import SeReMpy
+from SeReMpy.Inversion import *
+from SeReMpy.Geostats import CorrelatedSimulation
 
 #% Available data and parameters
-# Load data (seismic data, reference petroelastic properties, and time)
-ds = np.loadtxt('Data/data5seis.dat')
+# Load data (seismic data and time)
+ds = np.loadtxt('Data/data3seis.dat')
 TimeSeis = ds[:,0].reshape(-1, 1)
 Snear = ds[:,1].reshape(-1, 1)
 Smid = ds[:,2].reshape(-1, 1)
 Sfar = ds[:,3].reshape(-1, 1)
-dl = np.loadtxt('Data/data5log.dat')
-Phi = dl[:,0].reshape(-1, 1)
-Clay = dl[:,1].reshape(-1, 1)
-Sw = dl[:,2].reshape(-1, 1)
+dl = np.loadtxt('Data/data3log.dat')
+Vp = dl[:,0].reshape(-1, 1)
+Vs = dl[:,1].reshape(-1, 1)
+Rho = dl[:,2].reshape(-1, 1)
 Time = dl[:,3].reshape(-1, 1)
-Vp = dl[:,4].reshape(-1, 1)
-Vs = dl[:,5].reshape(-1, 1)
-Rho = dl[:,6].reshape(-1, 1)
+
 
 #% Initial parameters
 # number of samples (elastic properties)
@@ -61,13 +60,6 @@ freq = 45
 ntw = 64
 wavelet, tw = RickerWavelet(freq, dt, ntw)
 
-# matrix associated to the linear rock physics operator
-R = np.zeros((nd, nv + 1))
-X = np.hstack([Phi,Clay,Sw, np.ones(Phi.shape)])
-R[0, :] = (np.linalg.lstsq(X,Vp,rcond=None)[0]).T
-R[1, :] = (np.linalg.lstsq(X,Vs,rcond=None)[0]).T
-R[2, :] = (np.linalg.lstsq(X,Rho,rcond=None)[0]).T
-
 #% Plot seismic data
 plt.figure(1)
 plt.subplot(131)
@@ -92,10 +84,10 @@ plt.show()
 nfilt = 3
 cutofffr = 0.04
 b, a = signal.butter(nfilt, cutofffr)
-Phiprior = signal.filtfilt(b, a, np.squeeze(Phi))
-Clayprior = signal.filtfilt(b, a, np.squeeze(Clay))
-Swprior = signal.filtfilt(b, a, np.squeeze(Sw))
-mprior = np.hstack([Phiprior[:,np.newaxis], Clayprior[:,np.newaxis], Swprior[:,np.newaxis]])
+Vpprior = signal.filtfilt(b, a, np.squeeze(Vp))
+Vsprior = signal.filtfilt(b, a, np.squeeze(Vs))
+Rhoprior = signal.filtfilt(b, a, np.squeeze(Rho))
+mprior = np.hstack([Vpprior[:,np.newaxis], Vsprior[:,np.newaxis], Rhoprior[:,np.newaxis]])
 
 #% Spatial correlation matrix
 corrlength = 5 * dt
@@ -103,79 +95,64 @@ trow = np.matlib.repmat(np.arange(0, nm * dt, dt), nm, 1)
 tcol = np.matlib.repmat(trow[0,:].reshape(nm,1), 1, nm)
 tdis = abs(trow - tcol)
 sigmatime = np.exp(-(tdis / corrlength) ** 2)
-sigma0 = np.cov(np.hstack([Phi, Clay, Sw]).T)
+sigma0 = np.cov(np.hstack([Vp, Vs, Rho]).T)
 sigmaprior = np.kron(sigma0, sigmatime)
 
 #% Prior realizations
-Phisim = np.zeros((nm, nsim))
-Claysim = np.zeros((nm, nsim))
-Swsim = np.zeros((nm, nsim))
+Vpsim = np.zeros((nm, nsim))
+Vssim = np.zeros((nm, nsim))
+Rhosim = np.zeros((nm, nsim))
 SeisPred = np.zeros((nd * ntheta, nsim))
 for i in range(nsim):
     msim = CorrelatedSimulation(mprior, sigma0, sigmatime)
-    Phisim[:,i] = msim[:,0]
-    Claysim[:,i] = msim[:,1]
-    Swsim[:,i] = msim[:,2]
-Phisim[Phisim < 0] = 0
-Phisim[Phisim > 0.4] = 0.4
-Claysim[Claysim < 0] = 0
-Claysim[Claysim > 0.8] = 0.8
-Swsim[Swsim < 0] = 0
-Swsim[Swsim > 1] = 1
-Vpsim, Vssim, Rhosim = LinearizedRockPhysicsModel(Phisim, Claysim, Swsim, R)
-for i in range(nsim):
+    Vpsim[:,i] = msim[:,0]
+    Vssim[:,i] = msim[:,1]
+    Rhosim[:,i] = msim[:,2]
     seis, TimeSeis = SeismicModel(Vpsim[:,i], Vssim[:,i], Rhosim[:,i], Time, theta, wavelet)
     SeisPred[:,i] = seis[:,0]
-    
     
 # plot of prior models
 plt.figure(2)
 plt.subplot(131)
-plt.plot(Phisim, Time, 'b')
-plt.plot(Phi, Time, 'k')
-plt.plot(Phiprior, Time, 'r')
+plt.plot(Vpsim, Time, 'b')
+plt.plot(Vp, Time, 'k')
+plt.plot(Vpprior, Time, 'r')
 plt.grid()
 plt.ylim([max(Time), min(Time)])
-plt.xlabel('Porosity')
+plt.xlabel('P-wave velocity (km/s)')
 plt.ylabel('Time (s)')
 plt.subplot(132)
-plt.plot(Claysim, Time, 'b')
-plt.plot(Clay, Time, 'k')
-plt.plot(Clayprior, Time, 'r')
+plt.plot(Vssim, Time, 'b')
+plt.plot(Vs, Time, 'k')
+plt.plot(Vsprior, Time, 'r')
 plt.grid()
 plt.ylim([max(Time), min(Time)])
-plt.xlabel('Clay volume')
+plt.xlabel('S-wave velocity (km/s)')
 plt.subplot(133)
-plt.plot(Sw, Time, 'k')
-plt.plot(Swprior, Time, 'r')
-plt.plot(Swsim, Time, 'b')
-plt.plot(Sw, Time, 'k')
-plt.plot(Swprior, Time, 'r')
+plt.plot(Rho, Time, 'k')
+plt.plot(Rhoprior, Time, 'r')
+plt.plot(Rhosim, Time, 'b')
+plt.plot(Rho, Time, 'k')
+plt.plot(Rhoprior, Time, 'r')
 plt.grid()
 plt.ylim([max(Time), min(Time)])
-plt.xlabel('Water saturation')
+plt.xlabel('Density (g/cm^3)')
+plt.suptitle('Prior realizations')
 plt.legend(['Reference model', 'Prior mean','Prior Realizations'],loc ="lower right")
 plt.show()
 
 
-#% ESMDA petrophysical inversion
+#% ESMDA seismic inversion
 niter = 4
 alpha = 1 / niter   # sum alpha = 1
-PriorModels = np.vstack([Phisim, Claysim, Swsim])
+PriorModels = np.vstack([Vpsim, Vssim, Rhosim])
 SeisData = np.vstack([Snear,Smid,Sfar])
 PostModels = PriorModels
 for j in range(niter):
     PostModels, KalmanGain = EnsembleSmootherMDA(PostModels, SeisData, SeisPred, alpha, sigmaerr)
-    Phipost = PostModels[0:nm, :]
-    Claypost = PostModels[nm :2 * nm, :]
-    Swpost = PostModels[2 * nm :, :]
-    Phipost[Phipost < 0] = 0
-    Phipost[Phipost > 0.4] = 0.4
-    Claypost[Claypost < 0] = 0
-    Claypost[Claypost > 0.8] = 0.8
-    Swpost[Swpost < 0] = 0
-    Swpost[Swpost > 1] = 1
-    Vppost, Vspost, Rhopost = LinearizedRockPhysicsModel(Phipost, Claypost, Swpost, R)
+    Vppost = PostModels[0:nm, :]
+    Vspost = PostModels[nm :2 * nm, :]
+    Rhopost = PostModels[2 * nm :, :]
     for i in range(nsim):
         seis, TimeSeis = SeismicModel(Vppost[:,i], Vspost[:,i], Rhopost[:,i], Time, theta, wavelet)
         SeisPred[:,i] = seis[:,0]
@@ -183,36 +160,36 @@ for j in range(niter):
 # posterior mean models
 mpost = np.mean(PostModels, axis=1)
 mpost = mpost.reshape(len(mpost),1)
-Phimean = mpost[0:nm, 0]
-Claymean = mpost[nm :2 * nm, 0]
-Swmean = mpost[2 * nm :, 0]
+Vpmean = mpost[0:nm, 0]
+Vsmean = mpost[nm :2 * nm, 0]
+Rhomean = mpost[2 * nm :, 0]
 
 #% Plot results
 plt.figure(3)
 plt.subplot(131)
-plt.plot(Phipost, Time, 'b')
-plt.plot(Phi, Time, 'k')
-plt.plot(Phimean, Time, 'r')
+plt.plot(Vppost, Time, 'b')
+plt.plot(Vp, Time, 'k')
+plt.plot(Vpmean, Time, 'r')
 plt.grid()
 plt.ylim([max(Time), min(Time)])
-plt.xlabel('Porosity')
+plt.xlabel('P-wave velocity (km/s)')
 plt.ylabel('Time (s)')
 plt.subplot(132)
-plt.plot(Claypost, Time, 'b')
-plt.plot(Clay, Time, 'k')
-plt.plot(Claymean, Time, 'r')
+plt.plot(Vspost, Time, 'b')
+plt.plot(Vs, Time, 'k')
+plt.plot(Vsmean, Time, 'r')
 plt.grid()
 plt.ylim([max(Time), min(Time)])
-plt.xlabel('Clay volume')
+plt.xlabel('S-wave velocity (km/s)')
 plt.subplot(133)
-plt.plot(Sw, Time, 'k')
-plt.plot(Swmean, Time, 'r')
-plt.plot(Swpost, Time, 'b')
-plt.plot(Sw, Time, 'k')
-plt.plot(Swmean, Time, 'r')
+plt.plot(Rho, Time, 'k')
+plt.plot(Rhomean, Time, 'r')
+plt.plot(Rhopost, Time, 'b')
+plt.plot(Rho, Time, 'k')
+plt.plot(Rhomean, Time, 'r')
 plt.grid()
 plt.ylim([max(Time), min(Time)])
-plt.xlabel('Water saturation')
+plt.xlabel('Density (g/cm^3)')
+plt.suptitle('Posterior realizations')
 plt.legend(['Reference model', 'Posterior mean','Posterior Realizations'],loc ="lower right")
 plt.show()
-
