+++ Examples/GeostatisticsContinuousDriver.py
@@ -19,7 +19,9 @@ from scipy.io import loadmat
 import scipy.spatial.distance
 import matplotlib.pyplot as plt
 import numpy as np
-from Geostats import *
+
+from context import SeReMpy
+from SeReMpy.Geostats import *
 
 #% Example 1
 # available data (4 measurements)
