+++ Examples/context.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+import sys
+import os
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import SeReMpy
